Homework-4 v1
