+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\OpenServer20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\index.js":"1","C:\\OpenServer20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\App.js":"2","C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\index.js":"3","C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\App.js":"4","C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\components\\Layout\\index.js":"5","C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\components\\Header\\index.js":"6","C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\components\\Footer\\index.js":"7","C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\components\\PokemonCard\\index.js":"8"},{"size":214,"mtime":1611684244758,"results":"9","hashOfConfig":"10"},{"size":264,"mtime":1611684789543,"results":"11","hashOfConfig":"10"},{"size":146,"mtime":1611877228546,"results":"12","hashOfConfig":"13"},{"size":4077,"mtime":1612127699668,"results":"14","hashOfConfig":"13"},{"size":686,"mtime":1612125889820,"results":"15","hashOfConfig":"13"},{"size":413,"mtime":1612112449324,"results":"16","hashOfConfig":"13"},{"size":247,"mtime":1612025985759,"results":"17","hashOfConfig":"13"},{"size":1517,"mtime":1612127655274,"results":"18","hashOfConfig":"13"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"15u2obr",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"148mf5q",{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\OpenServer20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\index.js",[],"C:\\OpenServer20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\App.js",[],"C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\index.js",[],"C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\App.js",[],"C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\components\\Layout\\index.js",[],"C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\components\\Header\\index.js",[],"C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\components\\Footer\\index.js",[],"C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\components\\PokemonCard\\index.js",[]]
+[{"C:\\OpenServer20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\index.js":"1","C:\\OpenServer20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\App.js":"2","C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\index.js":"3","C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\App.js":"4","C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\components\\Layout\\index.js":"5","C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\components\\Header\\index.js":"6","C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\components\\Footer\\index.js":"7","C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\components\\PokemonCard\\index.js":"8"},{"size":214,"mtime":1611684244758,"results":"9","hashOfConfig":"10"},{"size":264,"mtime":1611684789543,"results":"11","hashOfConfig":"10"},{"size":151,"mtime":1612025985765,"results":"12","hashOfConfig":"13"},{"size":4077,"mtime":1612127699668,"results":"14","hashOfConfig":"13"},{"size":783,"mtime":1612218817355,"results":"15","hashOfConfig":"13"},{"size":413,"mtime":1612112449324,"results":"16","hashOfConfig":"13"},{"size":247,"mtime":1612025985759,"results":"17","hashOfConfig":"13"},{"size":1610,"mtime":1612220524561,"results":"18","hashOfConfig":"13"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"15u2obr",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"148mf5q",{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\OpenServer20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\index.js",[],"C:\\OpenServer20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\App.js",[],"C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\index.js",[],"C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\App.js",[],"C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\components\\Layout\\index.js",[],"C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\components\\Header\\index.js",[],"C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\components\\Footer\\index.js",[],"C:\\openserver20\\domains\\zar-react-marathon-2021\\elias-rwact-app\\src\\components\\PokemonCard\\index.js",[]]
