+++ src/components/Header/style.module.css
@@ -11,7 +11,7 @@
 .root::before {
 	content: '';
 	position: fixed;
-	background: url('../../assets/Stars.svg');
+	background: url('./assets/Stars.svg');
 	background-attachment: fixed;
 	width: 100%;
 	min-height: 100vh;
@@ -34,7 +34,7 @@
 	position: absolute;
 	bottom: -300px;
 	left: 0;
-	background: url('../../assets/Trees.svg') bottom left repeat-x;
+	background: url('./assets/Trees.svg') bottom left repeat-x;
 	background-size: cover;
 	width: 100%;
 	height: 80%;
